chore: reorganise users list on admin page
